[FAB2023-999] fix: 카테고리 데이터 추가 모달 - 코드리뷰 수정
